CARL ResourceManager: Move `persistence` module underneath `manager`.
