+++ opendut-carl/src/manager/cluster_manager/create_cluster_descriptor.rs
@@ -1,7 +1,7 @@
 use opendut_model::cluster::{ClusterDescriptor, ClusterId, ClusterName};
 use tracing::{debug, info};
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 
 pub struct CreateClusterDescriptorParams {
