+++ opendut-carl/src/manager/cluster_manager/delete_cluster_deployment.rs
@@ -4,7 +4,7 @@ use crate::settings::vpn::Vpn;
 use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment, ClusterId, ClusterName};
 use opendut_model::cluster::state::ClusterState;
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 
 pub struct DeleteClusterDeploymentParams {
