+++ opendut-carl/src/manager/cluster_manager/delete_cluster_descriptor.rs
@@ -3,7 +3,7 @@ use opendut_model::ShortName;
 use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment, ClusterId, ClusterName};
 use opendut_model::cluster::state::ClusterState;
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 
 pub struct DeleteClusterDescriptorParams {
