+++ opendut-carl/src/manager/cluster_manager/list_deployed_clusters.rs
@@ -1,4 +1,4 @@
-use crate::resource::persistence::error::{PersistenceError};
+use crate::resource::manager::error::{PersistenceError};
 use crate::resource::manager::ResourcesStorageApi;
 use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment};
 use std::collections::{HashSet};
