+++ opendut-carl/src/manager/observer_messaging_broker/observe_peers_online.rs
@@ -6,7 +6,7 @@ use opendut_carl_api::carl::observer::{WaitForPeersOnlineResponse, WaitForPeersO
 use opendut_model::peer::PeerId;
 use opendut_model::peer::state::PeerConnectionState;
 use crate::resource::manager::{ResourceManagerRef, SubscriptionEvent};
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::types::subscription::Subscription;
 
 pub struct ObservePeersOnline {
