+++ opendut-carl/src/manager/peer_manager/delete_peer_descriptor.rs
@@ -9,7 +9,7 @@ use opendut_model::cluster::ClusterId;
 use opendut_model::peer::state::{PeerMemberState, PeerState};
 use crate::manager::peer_manager::list_peer_member_states::ListPeerMemberStatesError;
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 
 pub struct DeletePeerDescriptorParams {
     pub vpn: Vpn,
