+++ opendut-carl/src/manager/peer_manager/generate_peer_setup.rs
@@ -1,4 +1,4 @@
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::settings::vpn::Vpn;
 use opendut_auth::registration::client::RegistrationClientRef;
 use opendut_auth::registration::resources::UserId;
