+++ opendut-carl/src/manager/peer_manager/get_peer_state.rs
@@ -1,5 +1,5 @@
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 use opendut_model::peer::state::{PeerConnectionState, PeerState};
 use opendut_model::peer::PeerId;
