+++ opendut-carl/src/manager/peer_manager/list_peer_member_states.rs
@@ -1,6 +1,6 @@
 use crate::manager::cluster_manager;
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 use opendut_model::cluster::ClusterId;
 use opendut_model::peer::state::PeerMemberState;
