+++ opendut-carl/src/manager/peer_manager/store_peer_descriptor.rs
@@ -1,5 +1,5 @@
 use crate::resource::types::resources::Resources;
-use crate::resource::persistence::error::PersistenceError;
+use crate::resource::manager::error::PersistenceError;
 use crate::resource::manager::ResourcesStorageApi;
 use crate::settings::vpn::Vpn;
 use opendut_model::peer::state::PeerState;
