+++ opendut-carl/src/resource/manager/persistence/storage/tests/cluster_descriptor.rs
