+++ opendut-carl/src/resource/mod.rs
@@ -1,3 +1,2 @@
 pub(crate) mod types;
 pub mod manager;
-pub mod persistence;
